refactor

+++ src/main/java/com/gted/x/controller/JobController.java
@@ -1,6 +1,7 @@
 package com.gted.x.controller;
 
 import com.gted.x.entity.*;
+import com.gted.x.exception.EntityNotFoundException;
 import com.gted.x.service.JobService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -17,12 +18,8 @@ import java.net.URL;
 public class JobController {
 
 
-    private JobService jobService;
-
     @Autowired
-    public JobController(JobService jobService) {
-        this.jobService = jobService;
-    }
+    private JobService jobService;
 
     @PostMapping(value = "/jobs")
     @SuppressWarnings("unchecked")
@@ -30,21 +27,25 @@ public class JobController {
 
         Job savedJob = jobService.add(requestJobBody.getJob());
         JobBody body = new JobBody(savedJob.getJobId());
-
         return new JobResponseBody<JobBody>(body);
     }
 
     @GetMapping(value = "/jobs/{job_id}")
     public JobResponseBody<Job> get(@PathVariable ("job_id") long id){
-
         Job job = jobService.getById(id);
+        if(job == null){
+            throw new EntityNotFoundException(String.format("job with %s not found", id));
+        }
         return new JobResponseBody<Job>(job);
     }
 
     @DeleteMapping(value = "/jobs/{job_id}")
     @SuppressWarnings("unchecked")
     public Response delete(@PathVariable ("job_id") long id){
-
+        Job job = jobService.getById(id);
+        if(job == null){
+            throw new EntityNotFoundException(String.format("job with ID %s not found", id));
+        }
         jobService.delete(id);
         return new Response(200, "jod deleted successfully");
     }
